feat: refactoring types
