+++ types/audit.ts
@@ -1,0 +1,5 @@
+
+type AuditType = {
+    label: string;
+    value: string;
+};
