+++ types/keyword.ts
@@ -1,0 +1,1 @@
+type Keyword = Record<"value" | "label", string>;
