+++ types/offer.ts
@@ -1,0 +1,14 @@
+type Offer = {
+    participants: string[];
+    id: number,
+    url: string,
+    badges: string[],
+    description: string,
+    budget: string,
+    from: string,
+    to: string,
+    type: string,
+    owner:string,
+    lines_count: number,
+    files_count: number,
+}
